[WIP] Add cluster-etcd plugin
This adds a first implementation of a plugin that watches for a node
config in etcd and pushes the results as cluster state to the local
node. I haven't tested this yet -- just wrote the code.

I also modified IndicesClusterStateService to let a local shard move
from INITIALIZING to STARTED without needing to go through a cluster
manager.

Signed-off-by: Michael Froh <msfroh@apache.org>