[WIP] Read node key on startup
When I added the node key watcher on startup, I assumed it would read
the current state as well. It turns out that it doesn't -- it only
waits for subsequent changes to the entry. Now we explicitly read.

Also, removed the getIndices method from NodeState, because I don't
think it's needed. I intended to use that to establish watchers for
index metadata changes, but I think I'm going to find a better way.

Signed-off-by: Michael Froh <msfroh@apache.org>