Set index to "search only" on search replica
Because of how "cluster health" (red, yellow, green) is calculated, we
normally consider it disastrous if a shard has no primary in the routing
table. In our clusterless environment, though, it's totally normal for
a search replica to live on a node with no reference to the primary in
the routing table.

In order to prevent core from panicking, we need to claim that the index
is running in "search only" mode, which means that it's okay to have no
primary shards.

Signed-off-by: Michael Froh <msfroh@apache.org>