Add support for search unit groups
For coordinator nodes in particular, we'd like to be able to support a
notion of "search unit groups". That is, we can publish a goal-state to
a path like:

/cluster-name/coordinators/default-coordinator/goal-state

Then all coordinators can read that same goal state, based on
attributes search_unit_group=coordinators,
search_unit=default-coordinator.

Similarly, we can use the search_unit_group to publish the
coordinator's heartbeats to:

/cluster-name/coordinators/node-name/actual-state

Signed-off-by: Michael Froh <msfroh@apache.org>