Add basic support for document replication
This is a pretty big change, since we need to propagate a lot more
information around allocation IDs.

I've been able to verify that this works on a two-node setup by
applying the following goal-states, in order, with enough time to
heartbeat between each step:

1. Node 0 has a primary, with no replicas.
2. Node 1 has a replica, pointing to node 0 for the primary.
3. Node 0 still has a primary, but now has node 1 as a replica.
4. Resubmit the state from step 3. (So node 0 sees that the replica
   has started.)

After that, it behaves like a regular primary/replica pair, where
either node is able to accept indexing requests and serve search
requests.

Signed-off-by: Michael Froh <msfroh@apache.org>