Fix cluster state updates that keep a shard around
Previous testing involved adding and removing a single shard from a
node. Unfortunately, we'll often add "another" shard to a node, which
means that the cluster state needs to honor attributes of existing
shards to avoid violating core assumptions.

This means that:
1. mappings must be byte-for-byte identical across updates (unless
   we bump the mapping version), and
2. shard allocation IDs must be maintained.

This change addresses the first by writing the mappings as
CompressedXContent, which preserves order. It addresses the second
by reusing an existing ShardRouting object if one existed in the
previous cluster state.

Signed-off-by: Michael Froh <msfroh@apache.org>