Retry loading goal state until cluster state converges (#26)
When we rely on status of shards on other nodes to apply goal state,
as is the case for document replication, what should we do when those
shards don't have the given status?

One option (implemented previously) is to throw an exception, declaring
the goal state invalid. Another option (implemented here) is to apply
the parts of goal state that we can, and say that we have not converged
to the desired goal state. When that happens, we can schedule a reload
of the goal state some time later, hoping that the other nodes have
made enough progress to allow us to converge.

Signed-off-by: Michael Froh <msfroh@apache.org>