Block cluster manager APIs (or route to local)
Since we don't have cluster managers on clusterless configurations, we
should block any mutating request that tries to route to the cluster
manager, since that's preferable to hanging on trying to reach the
cluster manager.

For read-only requests, we can force the operation to run on the local
node, which will usually read from the node's local cluster state.

Signed-off-by: Michael Froh <msfroh@apache.org>