Watch other etcd keys that may affect cluster state (#42)
In addition to watching the node's goal state, we should watch the
settings and mappings for any index with shards on the local node.
If we see a change to settings or mappings, we need to remember to
increment their version when applying them.

As a bonus, we can remove the polling logic for goal state convergence,
since we can add watches for the actual states of nodes that have not
yet converged.

Signed-off-by: Michael Froh <msfroh@apache.org>