Use MapperService to update mappings (#55)
Updating mappings in a way that satisfies the assertions in
MapperService is really hard. Cluster managers achieve it by creating
a new MapperService based on the old IndexMetadata and merging in the
new mapping. The resulting DocumentMapper is used as the source of
truth. This change aims to do the same thing in a few lines in
DataNodeState.

While in the code, I also took the time to move some OpenSearch-specific
logic out of ETCDStateDeserializer and into the NodeState subclasses.

Signed-off-by: Michael Froh <msfroh@apache.org>